feat: add loading overlay for pages that has service call (#152)
* feat: add loading overlay for pages that has service call

* fix: make overlay clickable so the components underneath is not clickable

+++ app/src/main/java/com/cmput301f20t21/bookfriends/ui/library/owned/OwnedViewModel.java
@@ -124,7 +124,9 @@ public class OwnedViewModel extends ViewModel {
         bookRepository.delete(book.getId())
                 .addOnSuccessListener(
                         res -> {
-                            bookRepository.deleteImage(book.getImageUrl());
+                            if (book.getImageUrl() != null) {
+                                bookRepository.deleteImage(book.getImageUrl());
+                            }
                             bookData.remove(book);
                             books.setValue(bookData);
                         })
